v2

+++ exams/views.py
@@ -24,13 +24,12 @@ from django.views.decorators.http import require_POST
 
 # Create your views here.
 
-
 class ExamList(generics.ListCreateAPIView):
     queryset = Exam.objects.all()
     serializer_class = ExamSerializer
 
-    # authentication_classes = [BasicAuthentication]
-    # authentication_classes = [TokenAuthentication]
+    authentication_classes = [BasicAuthentication]
+    authentication_classes = [TokenAuthentication]
     # permission_classes = [IsTeacherOrReadOnly]
 
 
@@ -78,8 +77,8 @@ class QuestionDetail(generics.RetrieveUpdateDestroyAPIView):
     authentication_classes = [TokenAuthentication]
     permission_classes = [IsTeacherOrReadOnly]
 
-    # authentication_classes = [TokenAuthentication]   
-    # permission_classes = [IsTeacherOrReadOnly]
+    authentication_classes = [TokenAuthentication]   
+    permission_classes = [IsTeacherOrReadOnly]
 
 
 class StudentAnswerListCreateView(generics.ListCreateAPIView):
@@ -101,9 +100,11 @@ class ExamAttemptRetrieveUpdateDestroyView(generics.RetrieveUpdateDestroyAPIView
     queryset = ExamAttempt.objects.all()
     serializer_class = ExamAttemptSerializer
     
-    
+import EMS
 
 def exam(request, pk):
+    if request.user.is_anonymous == True:
+        return redirect(EMS.views.signin)
     
     #  res = CostestQuestion.objects.filter(contest_id=pk)
     #     questions = [q.question_id for q in res]
@@ -114,18 +115,28 @@ def exam(request, pk):
     return render(request, 'exams/exam.html', {'questions': questions, "contest_id": pk, "subject": exam.subject})
     # return redirect(go_to_singin)
 def exams(request):
+    if request.user.is_anonymous == True:
+        return redirect(EMS.views.signin)
     
-    contest = Exam.objects.all()
-    return render(request, 'exams/exams.html', {'contests': contest})
+    is_teacher = request.user.groups.filter(name='Teacher_Group').exists()
+    is_manager = request.user.groups.filter(name='Principle').exists()
+    
+    contests = Exam.objects.all()
+    return render(request, 'exams/exams.html', {'contests': contests, "manger": is_manager, "is_teacher": is_teacher})
     
     messages.info(request, 'you must Login')
     return redirect(go_to_singin)
 
 def addexam(request):
+    if request.user.is_anonymous == True:
+        return redirect(EMS.views.signin)
+    
     return render(request, 'exams/add_exam.html', {})
 
 
 def editexam(request, pk):
+    if request.user.is_anonymous == True:
+        return redirect(EMS.views.signin)
     
     exam = Exam.objects.get(pk=pk)
     exam = ExamSerializer(exam).data
@@ -135,6 +146,9 @@ def editexam(request, pk):
 
 require_POST
 def deletexam(request, pk):
+    if request.user.is_anonymous == True:
+        return redirect(EMS.views.signin)
+    
     exam = get_object_or_404(Exam, pk=pk)
     exam.delete()
     return redirect(exams)
